pyproject.toml
[project]
name = "sdiff"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "diffusers>=0.32.2",
    "huggingface-hub>=0.29.1",
    "kornia>=0.8.0",
    "librosa>=0.10.2.post1",
    "matplotlib>=3.7.5",
    "moviepy>=1.0.3",
    "numpy>=1.24.4",
    "onnx>=1.17.0",
    "onnxruntime>=1.20.1",
    "opencv-python>=4.11.0.86",
    "pillow==9.5.0",
    "pycocotools>=2.0.8",
    "sam2>=1.1.0",
    "scipy>=1.10.1",
    "timm>=1.0.15",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
]

[tool.uv.workspace]
members = ["segmentation"]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
